+++ doc/changes.rst
@@ -10,6 +10,7 @@ Change Log
   variables.
 * Use :class:`distutils.command.sdist.sdist` to ensure that ``MANIFEST.in``
   is respected.
+* Add some test coverage in :mod:`desiutil.setup`.
 
 1.4.0 (2016-04-28)
 ------------------